refactor: 게시판 ID로 글 조회는 board 라우트로 변경
