+++ server/src/controllers/boardController/index.ts
@@ -1,0 +1,3 @@
+export { createBoardHandler } from "./createBoard";
+export { getBoardList } from "./getBoardList";
+export { getPostListByBoard } from "./getPostListByBoard";
