+++ server/src/controllers/postController/index.ts
@@ -1,5 +1,4 @@
 export { createPostHandler } from './createPost';
-export { getPostListByBoard } from './readPost';
 export { updatePostHandler } from './updatePost';
 export { deletePostHandler } from './deletePost';
 export { getPostByIdHandler } from './getPostById';
